Rename topologycache packge to topologyheuristics
