+++ staging/src/k8s.io/endpointslice/topologyheuristics/utils_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package topologycache
+package topologyheuristics
 
 import (
 	"reflect"
